implement new HTTP handling logic
for the proxy to work properly, the buffer size is set to 1,
otherwise g_io_read_chars() will accumlate data in the memory and
does not return. This is different from recv(), recv will return even
if the data received now can not fill full the buffer.

Maybe using unblocking read can solve the problem.